Convert utils to Typescript
